+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * To identify broken Microsoft Exchange messages, check if the headers include 'X-MS-Exchange' instead of specific domain names. Before this, we've missed mails sent by Exchange installations not operated by Microsoft or mails with a different "originating organisation domain" than Hotmail or Outlook. (#333)
 * Do not anymore fail on emails containing any PGP boundaries as part of their plain text. As a sideeffect we will not anymore validate an email a second time. Hence, a message part containing an additional signature within an encrypted (and possibly signed) email won't be validated and removed. (#261)
 * Exit with code 1 if a CLI-subcommand was not found (#339).
+* Fix finding keywords in request-messages that were sent from Thunderbird/Enigmail with enabled "protected subject".
+* Fix leaking the "protected subject" sent from Thunderbird/Enigmail.
 
 ### Known issues